Added blink and button press.
I've adde a blink function which will start or stop blink an LED. This
is a Particle function so we can call it over the network.

Also I added code to count button presses, and register this count as a
Particle variable so we can read the current value over the network.

Currently the system mode is set as semi-auto, since this lets us run
code before connecting to the cloud and supposedly reduces data usage.
We should compare vs auto mode to see if we end up connecting more,
since if we reconnect to the cell network too much we could get our SIM
banned :(