Improved connection code be able to receive data.
-The code now looks for the appropraite characteristic so it can receive
data.
-There is a callback for received data, which just prints the data for
now.

I can't figure out how to change the MTU. I'm not sure if this is
supposed to be done by the central or peripheral device.
The itsy bitsy does supposedly try to set it  to 247, but when I read it
later both devices report an  MTU of 23, which will not be optimal.