Started setup of command and telemtry BLE streams.
Two new charactersitics were added to the sensor node service.
These allow us to receive a telemetry stream, and send commands to the
sensor nodes using the command stream.

So this commit includes changes to subscribe the gateway to the
characteristics.

This can successfully connect to a device and subscribe to the three
characteristics. Actually using the new streams has not been tested.