Fix hardfault from connecting to node that is transmitting.
The boron would crash with hard fault error if it connects to a sensor
node that is transmitting data(so a reconnect scenario).

Problem is that during the connection sequence we subscribe to the
characteristics(data streams) before we update the vairable saying that
we have connected and assigning node index.
So the data arrives and in the callback we reference an index that has
not been updated (-1) so we try and index and array using -1 which
causes hard fault.

So to fix this we just move the characteristic subscription to after we
assign the index and add the connection.

Tested a few times and no more hard faults!